+++ src/impulse/dotfile.py
@@ -1,4 +1,28 @@
 from textwrap import dedent
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True, order=True)
+class Edge:
+    source: str
+    destination: str
+    label: str = ""
+    emphasized: bool = False
+
+    def __str__(self) -> str:
+        return f'"{DotGraph.render_module(self.source)}" ->  "{DotGraph.render_module(self.destination)}"{self._render_attrs()}\n'
+
+    def _render_attrs(self) -> str:
+        attrs: dict[str, str] = {}
+        if self.label:
+            attrs["label"] = self.label
+        if self.emphasized:
+            attrs["style"] = "dashed"
+        if attrs:
+            joined_attrs = ", ".join([f'{key}="{value}"' for key, value in attrs.items()])
+            return f" [{joined_attrs}]"
+        else:
+            return ""
 
 
 class DotGraph:
@@ -8,37 +32,34 @@ class DotGraph:
     https://en.wikipedia.org/wiki/DOT_(graph_description_language)
     """
 
-    def __init__(self, title: str) -> None:
+    def __init__(self, title: str, concentrate: bool = True) -> None:
         self.title = title
         self.nodes: set[str] = set()
-        self.edges: set[tuple[str, str, str]] = set()
+        self.edges: set[Edge] = set()
+        self.concentrate = concentrate
 
     def add_node(self, name: str) -> None:
         self.nodes.add(name)
 
-    def add_edge(self, *, source: str, destination: str, label: str) -> None:
-        self.edges.add((source, destination, label))
+    def add_edge(self, edge: Edge) -> None:
+        self.edges.add(edge)
 
     def render(self) -> str:
         # concentrate=true means that we merge the lines together.
         return dedent(f"""digraph {{
             node [fontname=helvetica]
-            concentrate=true
+            {"concentrate=true" if self.concentrate else ""}
             {self._render_nodes()}
             {self._render_edges()}
         }}""")
 
     def _render_nodes(self) -> str:
-        return "\n".join(f'"{self._render_module(node)}"\n' for node in sorted(self.nodes))
+        return "\n".join(f'"{self.render_module(node)}"\n' for node in sorted(self.nodes))
 
     def _render_edges(self) -> str:
-        return "\n".join(
-            f'"{self._render_module(source)}" ->  "{self._render_module(destination)}"{self._render_label(label)}\n'
-            for source, destination, label in sorted(self.edges)
-        )
+        return "\n".join(str(edge) for edge in sorted(self.edges))
 
-    def _render_module(self, module: str) -> str:
-        return module.rsplit(self.title)[1]
-
-    def _render_label(self, label: str) -> str:
-        return f" [label={label}]" if label else ""
+    @staticmethod
+    def render_module(module: str) -> str:
+        # Render as relative module.
+        return f".{module.split('.')[-1]}"