+++ tests/unit/application/test_use_cases.py
@@ -1,8 +1,9 @@
+from typing import Optional
 from impulse.application import use_cases
 from copy import copy
 from impulse import dotfile
+from impulse.dotfile import Edge
 import grimp
-from grimp.adaptors.graph import ImportGraph
 from impulse import ports
 
 SOME_ROOT_PACKAGE = "mypackage"
@@ -10,7 +11,7 @@ SOME_MODULE = f"{SOME_ROOT_PACKAGE}.foo"
 
 
 def build_fake_graph(package_name: str) -> grimp.ImportGraph:
-    graph = ImportGraph()
+    graph = grimp.ImportGraph()
     graph.add_module(package_name)
 
     graph.add_module(SOME_MODULE)
@@ -43,13 +44,18 @@ def build_fake_graph(package_name: str) -> grimp.ImportGraph:
         importer=f"{SOME_MODULE}.blue.delta",
         imported=f"{SOME_MODULE}.red.epsilon",
     )
+    # Add a cycle.
+    graph.add_import(
+        importer=f"{SOME_MODULE}.red.epsilon",
+        imported=f"{SOME_MODULE}.blue.alpha",
+    )
 
     return graph
 
 
 class SpyGraphViewer(ports.GraphViewer):
     def __init__(self) -> None:
-        self.called_with_dot: dotfile.DotGraph | None = None
+        self.called_with_dot: Optional[dotfile.DotGraph] = None
 
     def view(self, dot: dotfile.DotGraph) -> None:
         self.called_with_dot = dot
@@ -65,6 +71,7 @@ class TestDrawGraph:
         use_cases.draw_graph(
             SOME_MODULE,
             show_import_totals=False,
+            show_cycle_breakers=False,
             sys_path=sys_path,
             current_directory=current_directory,
             build_graph=build_fake_graph,
@@ -76,6 +83,7 @@ class TestDrawGraph:
         # The image generation function was called.
         assert viewer.called_with_dot, "Viewer not called."
         assert viewer.called_with_dot.title == SOME_MODULE
+        assert viewer.called_with_dot.concentrate is True
         assert viewer.called_with_dot.nodes == {
             "mypackage.foo.green",
             "mypackage.foo.blue",
@@ -83,9 +91,10 @@ class TestDrawGraph:
             "mypackage.foo.red",
         }
         assert viewer.called_with_dot.edges == {
-            ("mypackage.foo.blue", "mypackage.foo.green", ""),
-            ("mypackage.foo.green", "mypackage.foo.yellow", ""),
-            ("mypackage.foo.blue", "mypackage.foo.red", ""),
+            Edge("mypackage.foo.blue", "mypackage.foo.green"),
+            Edge("mypackage.foo.green", "mypackage.foo.yellow"),
+            Edge("mypackage.foo.blue", "mypackage.foo.red"),
+            Edge("mypackage.foo.red", "mypackage.foo.blue"),
         }
 
     def test_draw_graph_show_import_totals(self):
@@ -94,21 +103,47 @@ class TestDrawGraph:
         use_cases.draw_graph(
             SOME_MODULE,
             show_import_totals=True,
+            show_cycle_breakers=False,
             sys_path=[],
             current_directory="/cwd",
             build_graph=build_fake_graph,
             viewer=viewer,
         )
 
-        assert viewer.called_with_dot.title == SOME_MODULE
-        assert viewer.called_with_dot.nodes == {
-            "mypackage.foo.green",
-            "mypackage.foo.blue",
-            "mypackage.foo.yellow",
-            "mypackage.foo.red",
-        }
+        assert viewer.called_with_dot.concentrate is False
         assert viewer.called_with_dot.edges == {
-            ("mypackage.foo.blue", "mypackage.foo.green", "1"),
-            ("mypackage.foo.green", "mypackage.foo.yellow", "1"),
-            ("mypackage.foo.blue", "mypackage.foo.red", "4"),
+            Edge("mypackage.foo.blue", "mypackage.foo.green", label="1"),
+            Edge("mypackage.foo.green", "mypackage.foo.yellow", label="1"),
+            Edge("mypackage.foo.blue", "mypackage.foo.red", label="4"),
+            Edge("mypackage.foo.red", "mypackage.foo.blue", label="1"),
+        }
+
+    def test_draw_graph_show_cycle_breakers(self):
+        viewer = SpyGraphViewer()
+
+        use_cases.draw_graph(
+            SOME_MODULE,
+            show_import_totals=False,
+            show_cycle_breakers=True,
+            sys_path=[],
+            current_directory="/cwd",
+            build_graph=build_fake_graph,
+            viewer=viewer,
+        )
+
+        assert viewer.called_with_dot.concentrate is False
+        assert viewer.called_with_dot.edges == {
+            Edge(
+                "mypackage.foo.blue",
+                "mypackage.foo.green",
+            ),
+            Edge(
+                "mypackage.foo.green",
+                "mypackage.foo.yellow",
+            ),
+            Edge(
+                "mypackage.foo.blue",
+                "mypackage.foo.red",
+            ),
+            Edge("mypackage.foo.red", "mypackage.foo.blue", emphasized=True),
         }